feat(worker): move forwarding request and dead-letter into kernel
Keep worker paths as thin re-exports so existing tests keep importing the same modules while cores live under lib/kernel/worker.
