+++ lib/kernel/worker/ForwardingRequestError.js
@@ -1,0 +1,8 @@
+import { BaseError } from 'make-error';
+
+export default class ForwardingRequestError extends BaseError {
+  constructor(errorMessage, messageBody) {
+    super(errorMessage);
+    this.messageBody = messageBody;
+  }
+}
