+++ worker/src/handlers/statement/statementForwarding/statementForwardingDeadLetterHandler.js
@@ -1,18 +1,4 @@
-import Statement from 'lib/models/statement';
-
-const statementForwardingDeadLetterHandler = ({
-  statement,
-  statementForwarding
-}, done) => {
-  Statement.updateOne({ _id: statement._id }, {
-    $addToSet: {
-      deadForwardingQueue: statementForwarding._id
-    },
-    $pull: {
-      pendingForwardingQueue: statementForwarding._id
-    }
-  }).then(() => done())
-  .catch(err => done(err));
-};
-
-export default statementForwardingDeadLetterHandler;
+export {
+  default,
+  statementForwardingDeadLetterHandler,
+} from 'lib/kernel/worker/statementForwardingDeadLetter';
